wip: add test coverage for exception code path
